Extract shared LLM tool registry into core/llm_tools
The MCP server and the in-app AI chat were each maintaining parallel
tool-adapter layers: their own Tool primitive, their own wire-format
renderers, and their own dispatch (MCP's handler map vs the executor's
46-branch if/elif). The business logic was already shared (knowledge
Commands) — only the adapter machinery was duplicated.

Add core/llm_tools with one primitive (Tool, ToolContext, ToolError,
ToolRegistry) plus renderers (to_anthropic / to_openai / to_mcp). Each
consumer keeps its own curated tool set and registry instance but builds
on the shared primitive:

- ai_chat: notes_tools joins its JSON schemas with handlers (moved to
  notes_handlers) into NOTES_REGISTRY; NotesToolExecutor becomes a thin
  context-builder over the registry, keeping its public interface;
  anthropic_notes_tools / openai_notes_tools delegate to the shared
  renderers.
- mcp_server: tools.py builds a ToolRegistry; the view renders via
  to_mcp and dispatches through a ToolContext.

No behavior change — the two tool surfaces stay intentionally distinct.
